- Added new hooks to share logic between file upload components and the submission history components - Added paginated history to abstract the submission history table for the submission history component - Re-organized the submission history directory
