+++ tdrs-frontend/src/components/SubmissionHistory/components/PaginatedHistory.jsx
@@ -1,0 +1,45 @@
+import { useState } from 'react'
+import Paginator from '../../Paginator'
+
+export const PaginatedHistory = ({
+  caption,
+  files,
+  reprocessedState,
+  TableComponent,
+}) => {
+  const pageSize = 5
+  const [resultsPage, setResultsPage] = useState(1)
+
+  const pages =
+    files && files.length > pageSize ? Math.ceil(files.length / pageSize) : 1
+  const pageStart = (resultsPage - 1) * pageSize
+  const pageEnd = Math.min(files.length, pageStart + pageSize)
+
+  return (
+    <div
+      className="submission-history-section usa-table-container--scrollable"
+      style={{ maxWidth: '100%' }}
+      tabIndex={0}
+    >
+      <table className="usa-table usa-table--striped">
+        <caption>{caption}</caption>
+        {files && files.length > 0 ? (
+          <TableComponent
+            files={files.slice(pageStart, pageEnd)}
+            reprocessedState={reprocessedState}
+          />
+        ) : (
+          <span>No data available.</span>
+        )}
+      </table>
+
+      {pages > 1 && (
+        <Paginator
+          onChange={(page) => setResultsPage(page)}
+          selected={resultsPage}
+          pages={pages}
+        />
+      )}
+    </div>
+  )
+}
